Random number generator now works with size_t
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@870 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/Utils/RandomNumberGenerator.hh
@@ -99,7 +99,7 @@ private:
   size_t iterations_;
 
   /// maximum random number generated, which is used for normalization
-  double maxNum_;
+  size_t maxNum_;
 };
 
 //=============================================================================
